Update G1 robot USD path (#134)
## Summary
The G1 robot is already on the Isaac Nucleus drive that is synced to S3.
This updates the USD path from the hardcoded Nucleus one (which is only
accessible with NV login), to one using Isaac Lab's carb based directory
path, enabling automatically pulling the publicly accessible S3 version
of the asset.