Add section setting the API tokens to the documentation (#184)
## Summary
For now all of our assets are hosted on the internal nucleus server for
which an API token is needed. SQA already ran into this issue. This will
likely happen to other first (internal) users as well.
Therefore add a section about generating the API token to the
documentation installation section until we host our assets publicly.