Task timeouts (#211)
## Summary
Adds an optional specification of timeout to tasks.

## Detailed description
- We had the issue that if a policy failed at a task in a way it got
stuck, it would not time-out (and therefore fail) until the default
time-out, which was set at 30s. This was typically longer that than the
evaluation time. In such a scenario the failure would not be registered
as the failure, and not affect the success rate. A policy could get 100%
success by succeeding once and then getting stuck for ~30s.
- Now each task allows optional specification of a time out.